+++ frontend/src/components/Search/Results.module.css
@@ -33,6 +33,13 @@
   --width-col-season: 50px;
 }
 
+.resultsMobileHeader {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 0 1rem;
+}
+
 @media (min-width: 1320px) {
   .resultsContainer {
     --width-col-code: 110px;
