+++ frontend/src/components/Search/Results.tsx
@@ -6,6 +6,7 @@ import { FixedSizeList, FixedSizeGrid } from 'react-window';
 import { useShallow } from 'zustand/react/shallow';
 import FloatingWorksheet from './FloatingWorksheet';
 import LastUpdated from './LastUpdated';
+import RandomButton from './RandomButton';
 import ResultsGridItem from './ResultsGridItem';
 import ResultsHeaders from './ResultsHeaders';
 import ResultsItem from './ResultsItem';
@@ -142,7 +143,12 @@ function Results({
           numResults={data?.length ?? 0}
         />
       )}
-      {isMobile && <LastUpdated />}
+      {isMobile && (
+        <div className={styles.resultsMobileHeader}>
+          <RandomButton />
+          <LastUpdated />
+        </div>
+      )}
       {resultsListing}
       <FloatingWorksheet />
     </div>
