+++ frontend/src/components/Typography.module.css
@@ -107,6 +107,6 @@
 
 @media (max-width: calc(1320px - 0.5px)) {
   .smallText {
-    font-size: 64%;
+    font-size: 70%;
   }
 }
